Ticket #145 - Create an advanced search page.
This creates a quicktabs instance which displays the Islandora search in
one tab, and a tag-aware-search iframe in the other. This quicktabs
instance is embedded into a page at the path `cwrc_search`.

Note the current iframe points to the Digital Echidna site for now,
until we get the tag-aware-search URL.